requirements.txt
asyncio==3.4.3
pyyaml==6.0.1
